requirements.txt
# MCP Audio Server Dependencies
mcp-server>=0.1.0
pydantic>=2.0.0

# Audio processing and playback
pyttsx3>=2.90
pygame>=2.5.0
playsound>=1.3.0

# Text-to-speech alternatives
gTTS>=2.4.0

# System audio control
pycaw>=20240210; sys_platform == "win32"
pulsectl>=23.5.2; sys_platform == "linux"

# Cross-platform audio
sounddevice>=0.4.6
soundfile>=0.12.1

# Utilities
requests>=2.31.0
pathlib>=1.0.1
